Create RussianDollEnvelopes_18.cpp
Key Takeaway:

Its a problem on LIS using the technique of Binary Search. 
Just maintaining the LIS array and push element to the array only if the current element is larger then the element present at  the back of the LIS array.
Otherwise , find the index where the current element fits using BS and put it over there. The length of the resultant array will give us the answer.